First GENe Commit
Usage of the Biopython library is up and running well.  Current problems
are that runtime on the queries are incredibly high and variable in
comparison to the direct queries at the NCBI BLAST website.  A query
using this program usually takes 17- 100 seconds, where as on the site
it takes about 2-4.   Not sure why this is - it seems to be something on
Biopython's end.

Implementation of decision algorithm was much easier than expected.