Fixed a few bugs.
I fixed the problem that made the user run the program as an administrator.  You shouldn't have to do that anymore. I also added a few error messages that will pop up, made the README more readable, made it so you don't have to worry about different types of quotes when typing into the local database section, and a few other small things. I would suggest downloading this version if you are planning on using it today.